check amount_out_min

+++ contracts/aggregator/src/models.rs
@@ -32,7 +32,7 @@ use soroban_sdk::{contracttype, Address, String, Vec};
 pub struct DexDistribution {
     pub protocol_id: String,
     pub path: Vec<Address>,
-    pub parts: i128,
+    pub parts: u32,
 }
 
 #[contracttype]
